Create smith number
smith number is  a type of number which number has equal sum of it's digit with  it's prime factor's sum of digit;  

example:                                                                                                                               
666=6+6+6=18                                                                                                                            
and                                                                                                                                           
   it's prime factors are: 2,3,3,37;                                                                                                          
   → 2+3+3+3+7=18;                                                                                                             
  so,666 is a smith number ;

hint:                                                                                                                                        
first of we have find the sum of the digit of numbers then we have to find the prime of the numbers and then sum them up. check if equal then we said that it's smith number. nor, not!